+++ src/main/java/com/sprint/mission/discodeit/auth/AdminAccountInitializer.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.common.init;
+package com.sprint.mission.discodeit.auth;
 
 import com.sprint.mission.discodeit.entity.Role;
 import com.sprint.mission.discodeit.entity.User;
