+++ src/main/java/com/sprint/mission/discodeit/auth/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.security.jwt;
+package com.sprint.mission.discodeit.auth.jwt;
 
 import org.springframework.stereotype.Component;
